chore: separate sidebar layout from api component
